requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
SQLAlchemy==2.0.40
APScheduler==3.11.0
pydantic-settings==2.9.1
python-dotenv==1.1.0
jinja2==3.1.6
httpx==0.28.1
cryptography==44.0.3
itsdangerous==2.2.0
